add roles validation in user

+++ src/modules/users/user.resolver.ts
@@ -1,7 +1,7 @@
 import { Resolver, Query, Args, Context, Mutation } from '@nestjs/graphql';
 import { UserService } from './user.service';
 import { UpdateUserInput } from './dto/update-user.input';
-import { UseGuards, UnauthorizedException, SetMetadata, Inject } from '@nestjs/common';
+import { UseGuards, UnauthorizedException, SetMetadata, Inject, HttpException, HttpStatus } from '@nestjs/common';
 import { GraphqlAuthGuard } from '../auth/guards/gql-auth.guard';
 import { CurrentUser } from '../../core/decorators/user.decorator';
 import { User } from '../../shared/types/user';
@@ -9,6 +9,7 @@ import { UserType } from './model/user.model';
 import { RolesGuard } from '../auth/guards/roles.guard';
 import { Roles } from '../../core/decorators/roles.decorator';
 import { UserRoles } from '../../shared/user-roles';
+import { validRole } from 'src/shared/helpers/helper-functions';
 
 @UseGuards(GraphqlAuthGuard)
 @UseGuards(RolesGuard)
@@ -16,32 +17,28 @@ import { UserRoles } from '../../shared/user-roles';
 export class UserResolver {
   constructor(
     private userService: UserService
-    ) {}
+  ) { }
 
-  // @Roles("Admin")
-  @Roles("employee")
+  // @Roles("super")
 
   @Query(returns => UserType)
   async getUser(@CurrentUser() user: any) {
-    console.log("CurrentUser " , user);
-      return await this.userService.getUsers();
+    return await this.userService.getUsers();
   }
 
-  // @Mutation(returns => UserType)
-  // async delete(@Args('email') email: string, @CurrentUser() user: User) {
-  //   if (email === user.email) {
-  //     return await this.userService.deleteUserByEmail(email);
-  //   } else {
-  //     throw new UnauthorizedException();
-  //   }
-  // }
-
   @Mutation(returns => UserType)
-  async createUser(@Args() args: any, @CurrentUser() user: any) {
-  // async createUser(@Args() args: any) {
-    const { createUser } = args;
-    const response = await this.userService.create(createUser);
-    return response;
+  async createUser(@Args('createUser') createUser: any, @CurrentUser() currentUser: any) {
+    if (createUser.userRole == undefined) {
+      createUser.userRole = UserRoles.EMPLOYEE;
+    }
+    const validateTole = await validRole(currentUser.userRole, createUser.userRole);
+    if (validateTole) {
+      const response = await this.userService.create(createUser);
+      return response;
+    }
+    else {
+      throw new HttpException(`${currentUser.userRole} can't create ${createUser.userRole}`, HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Mutation(returns => UserType)
@@ -58,13 +55,40 @@ export class UserResolver {
   }
 
   @Mutation(returns => UserType)
-  async deleteUser(@Args('email') email: string) {
-      return await this.userService.deleteUserByEmail(email);
+  async deleteUserByEmail(@Args('email') email: string,
+    @CurrentUser() currentUser: User) {
+    const user = await this.userService.getUsersByUserEmail(email);
+    if (user != null) {
+      const validateTole = await validRole(currentUser.userRole, user.userRole);
+      if (validateTole) {
+        return await this.userService.deleteUserByEmail(email);
+      }
+      else {
+        throw new HttpException(`${currentUser.userRole} can't delete ${user.userRole}`, HttpStatus.BAD_REQUEST);
+      }
+    }
+    else {
+      return user;
+    }
+
   }
 
   @Mutation(returns => UserType)
-  async deleteUserById(@Args("userId") userId :string){
-    return await this.userService.deleteUserById(userId);
+  async deleteUserById(@Args("userId") userId: string,
+    @CurrentUser() currentUser: User) {
+    const user = await this.userService.getUsersByUserId(userId);
+    const validateTole = await validRole(currentUser.userRole, user.userRole);
+    if (user != null) {
+      if (validateTole) {
+        return await this.userService.deleteUserById(userId);
+      }
+      else {
+        throw new HttpException(`${currentUser.userRole} can't delete ${user.userRole}`, HttpStatus.BAD_REQUEST);
+      }
+    }
+    else {
+      return user;
+    }
   }
 
   @Mutation(returns => UserType)
@@ -74,7 +98,7 @@ export class UserResolver {
 
   @Mutation(returns => UserType)
   async getUsersByGroupId(@Args('groupId') groupId: string) {
-      return await this.userService.getUsersByGroupId(groupId);
+    return await this.userService.getUsersByGroupId(groupId);
   }
 
 }
